requirements.txt
streamlit==1.55.0
torch==2.5.1
torchvision==0.20.1
Pillow==12.0.0
matplotlib==3.10.8
deep-translator==1.11.4
gdown==5.2.0